fix(shared,client,admin): clear CI guardrails
