+++ packages/admin/src/components/Action/ActionTranslationEditor.tsx
@@ -260,11 +260,9 @@ export function ActionTranslationEditor({
       data: updater(record.data),
     }));
   };
-
   const updateTopLevel = (key: "title" | "description", nextValue: string) => {
     updateActiveData((data) => ({ ...data, [key]: nextValue }));
   };
-
   const updateScopedField = (
     scope: TranslationScope,
     key: string,
@@ -277,7 +275,6 @@ export function ActionTranslationEditor({
       }))
     );
   };
-
   const updateInputField = (
     path: string[],
     updater: (input: ActionInstructionInputTranslation) => ActionInstructionInputTranslation
@@ -293,7 +290,6 @@ export function ActionTranslationEditor({
       }))
     );
   };
-
   const handleGenerateDraft = async () => {
     setMessage(null);
     if (!browserTranslator.isSupported) {
