+++ packages/client/src/components/Public/VaultCardWalletManage.tsx
@@ -8,14 +8,11 @@
  */
 import {
   type Address,
-  DEFAULT_WITHDRAW_MAX_LOSS_BPS,
   formatTokenAmount,
   getOctantVaultCampaignBySlug,
   getOctantVaultPendingFundedCardWalletRefs,
-  OCTANT_VAULT_REDEEM_CALL_SHAPES,
   type OctantVaultCardWalletPositionRef,
   type OctantVaultPosition,
-  type OctantVaultRedeemCallVariant,
   rememberOctantVaultCardWalletPosition,
   truncateAddress,
   useOctantVaultPositions,
@@ -39,6 +36,10 @@ import {
 } from "thirdweb/react";
 import { inAppWallet, preAuthenticate } from "thirdweb/wallets/in-app";
 import { EditorialGhostButton } from "./atoms";
+import {
+  prepareCardWalletRedeem,
+  readCardWalletMaxRedeemable,
+} from "./VaultCardWalletManage.calls";
 import { PositionsList, VaultPositionRowView } from "./VaultManagePositionsPanel";
 
 function getThirdwebClientId(): string {
@@ -47,84 +48,6 @@ function getThirdwebClientId(): string {
 
 function getThirdwebChain(chainId: number) {
   return chainId === ethereum.id ? ethereum : defineChain(chainId);
-}
-
-async function readCardWalletMaxRedeemable({
-  contract,
-  owner,
-}: {
-  contract: ReturnType<typeof getContract>;
-  owner: Address;
-}): Promise<{ shares: bigint; variant: OctantVaultRedeemCallVariant }> {
-  for (const shape of OCTANT_VAULT_REDEEM_CALL_SHAPES) {
-    try {
-      const result = await readCardWalletMaxRedeemVariant(contract, owner, shape.variant);
-      return {
-        shares: typeof result === "bigint" ? result : 0n,
-        variant: shape.variant,
-      };
-    } catch {
-      // Try the next Octant V2 / ERC-4626-compatible redeem shape.
-    }
-  }
-
-  return { shares: 0n, variant: "multistrategy" };
-}
-
-function readCardWalletMaxRedeemVariant(
-  contract: ReturnType<typeof getContract>,
-  owner: Address,
-  variant: OctantVaultRedeemCallVariant
-) {
-  if (variant === "multistrategy") {
-    return readContract({
-      contract,
-      method:
-        "function maxRedeem(address owner, uint256 maxLoss, address[] strategies) view returns (uint256)",
-      params: [owner, DEFAULT_WITHDRAW_MAX_LOSS_BPS, []] as const,
-    });
-  }
-  if (variant === "tokenized-strategy") {
-    return readContract({
-      contract,
-      method: "function maxRedeem(address owner, uint256 maxLoss) view returns (uint256)",
-      params: [owner, DEFAULT_WITHDRAW_MAX_LOSS_BPS] as const,
-    });
-  }
-  return readContract({
-    contract,
-    method: "function maxRedeem(address owner) view returns (uint256)",
-    params: [owner] as const,
-  });
-}
-
-function prepareCardWalletRedeem(
-  contract: ReturnType<typeof getContract>,
-  shares: bigint,
-  owner: Address,
-  variant: OctantVaultRedeemCallVariant
-) {
-  if (variant === "multistrategy") {
-    return prepareContractCall({
-      contract,
-      method:
-        "function redeem(uint256 shares, address receiver, address owner, uint256 maxLoss, address[] strategies) returns (uint256)",
-      params: [shares, owner, owner, DEFAULT_WITHDRAW_MAX_LOSS_BPS, []] as const,
-    });
-  }
-  if (variant === "tokenized-strategy") {
-    return prepareContractCall({
-      contract,
-      method:
-        "function redeem(uint256 shares, address receiver, address owner, uint256 maxLoss) returns (uint256)",
-      params: [shares, owner, owner, DEFAULT_WITHDRAW_MAX_LOSS_BPS] as const,
-    });
-  }
-  return prepareContractCall({
-    contract,
-    method: "function redeem(uint256 shares, address receiver, address owner) returns (uint256)",
-    params: [shares, owner, owner] as const,
-  });
 }
 
 // The recovered email wallet config must match the one Card Endow connected with,
